metrics2: refactoring
Refactoring by splitting the code from one file to multiple
modules.
